requirements.txt
beautifulsoup4==4.5.1
coolsms-python-sdk==2.0.1
Django==1.10.2
google-api-python-client==1.5.3
httplib2==0.9.2
oauth2client==3.0.0
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
requests==2.11.1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
